requirements/local.txt
# Local development dependencies
-r base.txt
django-debug-toolbar==2.2
